# api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Workload Platform Control Plane API
  version: 1.0.0
  description: Public control-plane API for versioned workflow definitions and reliable workflow runs.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/workflows:
    get:
      operationId: listWorkflows
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A keyset-paginated page of workflow summaries.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkflowPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      operationId: createWorkflow
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkflowDefinition' }
      responses:
        '201':
          description: The first immutable workflow version was created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DefinitionRef' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/workflows/{workflow-id}:
    get:
      operationId: getWorkflow
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      responses:
        '200':
          description: The bounded summary of a workflow and its latest version.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkflowSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/workflows/{workflow-id}/versions:
    get:
      operationId: listWorkflowVersions
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A keyset-paginated page of immutable workflow versions.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VersionPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      operationId: createWorkflowVersion
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkflowDefinition' }
      responses:
        '201':
          description: A new immutable workflow version was created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DefinitionRef' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/workflows/{workflow-id}/versions/{version}:
    get:
      operationId: getWorkflowVersion
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
        - $ref: '#/components/parameters/Version'
      responses:
        '200':
          description: The requested immutable workflow definition.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkflowDefinition' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/workflows/{workflow-id}/versions/{version}/runs:
    post:
      operationId: startRun
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
        - $ref: '#/components/parameters/Version'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '202':
          description: A pending Run was durably created and accepted for asynchronous execution.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StartRunResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/runs:
    get:
      operationId: listRuns
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: workflow_id
          in: query
          required: false
          description: Return only Runs created from this workflow.
          schema: { type: string }
        - name: status
          in: query
          required: false
          description: Return only Runs currently in this workflow status.
          schema: { $ref: '#/components/schemas/WorkflowStatus' }
      responses:
        '200':
          description: A filtered, keyset-paginated page of bounded Run summaries.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/runs/{run-id}:
    get:
      operationId: getRun
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RunID'
      responses:
        '200':
          description: The bounded summary of one Run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/runs/{run-id}/tasks:
    get:
      operationId: listTasks
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RunID'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A stable task-index page without Attempt history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/runs/{run-id}/tasks/{task-key}:
    get:
      operationId: getTask
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RunID'
        - $ref: '#/components/parameters/TaskKey'
      responses:
        '200':
          description: One task and its complete Attempt history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/runs/{run-id}/events:
    get:
      operationId: listEvents
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RunID'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A stable event-sequence page of persisted state transitions.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/runs/{run-id}/cancel:
    post:
      operationId: cancelRun
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RunID'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: The persisted cancellation request for the Run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CancelRunResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/workers/register:
    post:
      operationId: registerWorker
      security: [{ workerBootstrapAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterWorkerRequest' }
      responses:
        '201':
          description: A new Worker session was created. The session token is returned only in this response.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisterWorkerResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /api/v1/workers/{worker-id}/leases/{dispatch-id}/complete:
    post:
      operationId: completeWorkerLease
      security: [{ workerSessionAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/WorkerID'
        - $ref: '#/components/parameters/DispatchID'
        - $ref: '#/components/parameters/WorkerProtocolVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteRequest' }
      responses:
        '200':
          description: The result was accepted or safely replayed for the current lease.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompleteResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/WorkerConflict' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /api/v1/workers:
    get:
      operationId: listWorkers
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A keyset-paginated page of Worker summaries without credentials.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkerPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    workerBootstrapAuth:
      type: http
      scheme: bearer
      description: The separately configured bootstrap token accepted only by Worker registration.
    workerSessionAuth:
      type: http
      scheme: bearer
      description: The one-time session token returned by Worker registration.
  parameters:
    WorkflowID:
      name: workflow-id
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque keyset cursor returned by the same resource, scope, and filters.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return.
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  responses:
    BadRequest:
      description: The request syntax, parameters, cursor, or workflow definition are invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: The Bearer token is missing or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    InternalError:
      description: The server could not complete the operation because of an internal error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    WorkflowPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/WorkflowSummary' }
        next_cursor: { type: string }
...